Module lifeguard::settings

Structs

MaxSize
Specifies the largest number of values the Pool will hold before it will begin to drop values being returned to it.

StartingSize
Specifies how many values should be requested from the Supplier at initialization time. These values will be available for immediate use.

Supplier
Specifies a value implementing Supply<Output=T> that will be used to allocate new values. If unspecified, T::new() will be invoked.

Traits

OptionSetter
Implementing this trait allows a struct to act as a configuration parameter in the builder API.