Struct lifeguard::RcRecycled

pub struct RcRecycled<T> where T: Recycleable {
    // some fields omitted
}

A smartpointer which uses reference counting (Rc) to know when to move its wrapped value back to the Pool that issued it.

Methods

impl<'a, T> RcRecycled<T> where T: Recycleable

fn detach(self) -> T

Disassociates the value from the Pool that issued it. This destroys the smartpointer and returns the previously wrapped value.