Struct lifeguard::Recycled

pub struct Recycled<'a, T: 'a> where T: Recycleable { /* fields omitted */ }

A smartpointer which uses a shared reference (&) to know when to move its wrapped value back to the Pool that issued it.

Methods

impl<'a, T> Recycled<'a, T> where T: Recycleable

fn detach(self) -> T

Disassociates the value from the Pool that issued it. This destroys the smartpointer and returns the previously wrapped value.