Trait lifeguard::InitializeWith

pub trait InitializeWith<T> {
    fn initialize_with(&mut self, source: T);
}

Informs how an already allocated value should be initialized when provided with a model value or other meaningful input.

Required methods

fn initialize_with(&mut self, source: T)

Loading content...